+++ lib/avro/builder/errors.rb
@@ -11,5 +11,19 @@ module Avro
         super("attribute :#{attribute} missing for #{location}type :#{type}")
       end
     end
+
+    class DuplicateDefinitionError < StandardError
+      def initialize(key, object, existing_object)
+        super("definition for #{key.inspect} already exists\n"\
+              "existing definition:\n#{to_json(existing_object)}\n"\
+              "new definition:\n#{to_json(object)})")
+      end
+
+      private
+
+      def to_json(object)
+        object.to_h(SchemaSerializerReferenceState.new).to_json
+      end
+    end
   end
 end