+++ lib/avro/builder/field.rb
@@ -12,13 +12,11 @@ module Avro
 
       INTERNAL_ATTRIBUTES = Set.new(%i(optional_field)).freeze
 
-      attr_accessor :type, :optional_field, :builder, :record
-
       # These attributes may be set as options or via a block in the DSL
       dsl_attributes :doc, :aliases, :default, :order
 
-      def initialize(name:, type_name:, record:, builder:, internal: {}, options: {}, &block)
-        @builder = builder
+      def initialize(name:, type_name:, record:, cache:, internal: {}, options: {}, &block)
+        @cache = cache
         @record = record
         @name = name.to_s
 
@@ -33,15 +31,18 @@ module Avro
         @type = if builtin_type?(type_name)
                   create_and_configure_builtin_type(type_name,
                                                     field: self,
+                                                    cache: cache,
                                                     internal: internal,
                                                     options: options)
                 else
-                  builder.lookup_named_type(type_name)
+                  named_type = true
+                  cache.lookup_named_type(type_name)
                 end
 
         # DSL calls must be evaluated after the type has been constructed
         instance_eval(&block) if block_given?
         @type.validate!
+        @type.cache! unless named_type
       end
 
       ## Delegate additional DSL calls to the type
@@ -81,6 +82,8 @@ module Avro
       end
 
       private
+
+      attr_accessor :type, :optional_field, :cache, :record
 
       # Optional fields must be serialized as a union -- an array of types.
       def serialized_type(reference_state)