+++ lib/avro/builder/type_factory.rb
@@ -10,13 +10,13 @@ module Avro
       private
 
       # Return a new Type instance
-      def create_builtin_type(type_name)
+      def create_builtin_type(type_name, field:, cache:)
         name = type_name.to_s.downcase
         case
         when Avro::Schema::PRIMITIVE_TYPES.include?(name)
-          Avro::Builder::Types::Type.new(name)
+          Avro::Builder::Types::Type.new(name, field: field, cache: cache)
         when COMPLEX_TYPES.include?(name)
-          Avro::Builder::Types.const_get("#{name.capitalize}Type").new
+          Avro::Builder::Types.const_get("#{name.capitalize}Type").new(field: field, cache: cache)
         else
           raise "Invalid builtin type: #{type_name}"
         end
@@ -26,13 +26,11 @@ module Avro
       # and setting attributes via the DSL
       def create_and_configure_builtin_type(type_name,
                                             field: nil,
-                                            builder: nil,
+                                            cache: nil,
                                             internal: {},
                                             options: {},
                                             &block)
-        create_builtin_type(type_name).tap do |type|
-          type.field = field
-          type.builder = builder
+        create_builtin_type(type_name, field: field, cache: cache).tap do |type|
           type.configure_options(internal.merge(options))
           type.instance_eval(&block) if block_given?
         end