+++ lib/avro/builder/types/type_referencer.rb
@@ -8,15 +8,11 @@ module Avro
       module TypeReferencer
         include Avro::Builder::TypeFactory
 
-        def builder
-          (!field.nil? && field.builder) || super
-        end
-
         def create_builtin_or_lookup_named_type(type_name)
           if builtin_type?(type_name)
-            create_builtin_type(type_name)
+            create_builtin_type(type_name, field: field, cache: cache)
           else
-            builder.lookup_named_type(type_name)
+            cache.lookup_named_type(type_name)
           end
         end
       end